requirements.txt
anthropic==0.49.0
datasets==3.2.0
names-dataset==3.1.0
openai==1.75.0
PyYAML==6.0.2
tqdm==4.67.1
antlr4-python3-runtime==4.13.2
math-verify[antlr4_13_2]==0.6.1